+++ src/main/java/com/nhom7/coworkingspace/service/UserService.java
@@ -1,5 +1,6 @@
 package com.nhom7.coworkingspace.service;
 
+import com.nhom7.coworkingspace.dto.request.UpdateUserRequest;
 import com.nhom7.coworkingspace.dto.response.UpdateUserRoleResponse;
 import com.nhom7.coworkingspace.dto.response.UserProfileResponse;
 
@@ -10,4 +11,6 @@ public interface UserService {
     UpdateUserRoleResponse addRole(Long userId, String roleName);
 
     UserProfileResponse getMyProfile(String email);
+
+    UserProfileResponse updateMyProfile(String email, UpdateUserRequest request);
 }